requirements.txt
absl-py==0.15.0
astor==0.8.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.12
clang==5.0
click==8.0.4
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
dataclasses==0.8
Flask==2.0.3
flatbuffers==1.12
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==2.8.0
idna==3.10
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver @ file:///C:/ci/kiwisolver_1612282446297/work
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib @ file:///C:/ci/matplotlib-suite_1613408055530/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
nncase @ file:///D:/nncase/nncase-1.5.0.20220331-cp36-cp36m-win_amd64.whl
numpy==1.19.5
oauthlib==3.2.2
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.2.0.32
opt-einsum==3.3.0
pandas @ file:///C:/ci/pandas_1608056614942/work
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2021.3
PyYAML==6.0.1
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==0.24.2
scipy==1.4.1
seaborn @ file:///tmp/build/80754af9/seaborn_1629307859561/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.1.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
Theano==1.0.5
threadpoolctl==3.1.0
tornado @ file:///C:/ci/tornado_1606942379977/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
typing-extensions==3.7.4.3
urllib3==1.26.20
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0
